Use VariantTag/u16 in appropriate places

+++ external-crates/move/crates/move-core-types/src/compressed/runtime/serde_impl.rs
@@ -112,7 +112,7 @@ impl<'d> serde::de::Visitor<'d> for CompressedEnumFieldVisitor<'_> {
             None => return Err(A::Error::invalid_length(0, &self)),
         };
 
-        let variant_fv = match self.0.variant(tag as usize) {
+        let variant_fv = match self.0.variant(tag) {
             Some(VariantLayout::Known(fv)) => fv,
             Some(VariantLayout::Unknown) => {
                 return Err(A::Error::custom(format!(
